# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "岚山区劳动争议调解区块链平台，后台采用golang开发，使用gin + gorm + gorbac + jwt开发",
        "title": "岚山区劳动争议调解区块链平台",
        "termsOfService": "http://hafrans.com",
        "contact": {
            "name": "Chuuka Ro (Hafrans)",
            "url": "http://hafrans.com/support",
            "email": "lvzh@hafrans.com"
        },
        "license": {},
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auth/employer/register": {
            "post": {
                "description": "用人单位账户注册",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user",
                    "register"
                ],
                "summary": "用人单位 账户注册",
                "parameters": [
                    {
                        "description": "请求",
                        "name": "email",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.UserRegisterEmployerForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "正常业务处理",
                        "schema": {
                            "$ref": "#/definitions/vo.Common"
                        }
                    },
                    "401": {
                        "description": "未验证",
                        "schema": {
                            "$ref": "#/definitions/vo.Common"
                        }
                    },
                    "422": {
                        "description": "表单绑定失败",
                        "schema": {
                            "$ref": "#/definitions/vo.Common"
                        }
                    },
                    "500": {
                        "description": "表单绑定失败",
                        "schema": {
                            "$ref": "#/definitions/vo.Common"
                        }
                    }
                }
            }
        },
...
        "/api/auth/login": {
            "post": {
                "description": "用户登录并获得token以及失效日期",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "手机号",
                        "name": "phone",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "密码",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "验证码",
                        "name": "captcha_code",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "验证码时间戳 （2006-01-02 15:04:05）",
                        "name": "captcha_time",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "验证码挑战指令",
                        "name": "captcha_challenge",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/vo.LoginResult"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/vo.Common"
                        }
                    }
                }
            }
        },
...
        "/api/v1/record/create": {
            "post": {
                "description": "由管理员或者部门人员创建记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "case",
                    "record"
                ],
                "summary": "根据 CaseID 创建记录",
                "parameters": [
                    {
                        "description": "提交表单",
                        "name": "case",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.Record"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/vo.CommonData"
                        }
                    },
                    "401": {
                        "description": "没有认证",
                        "schema": {
                            "$ref": "#/definitions/vo.Common"
                        }
                    },
                    "422": {
                        "description": "绑定失败",
                        "schema": {
                            "$ref": "#/definitions/vo.Common"
                        }
                    }
                }
            }
        },
        "/api/v1/record/delete/:id": {
            "get": {
                "description": "根据record 的id号码删除record，只有管理员、部门人员可以操作",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "labor",
                    "record"
                ],
                "summary": "根据record 的id号码删除record",
                "parameters": [
                    {
                        "type": "number",
                        "description": "表单id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "正常业务处理",
                        "schema": {
                            "$ref": "#/definitions/vo.CommonData"
                        }
                    },
                    "401": {
                        "description": "未验证",
                        "schema": {
                            "$ref": "#/definitions/vo.Common"
                        }
                    }
                }
            }
        },
        "/api/v1/suggestion/create": {
            "post": {
                "description": "由管理员或者部门人员创建部门建议",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "case",
                    "suggestion"
                ],
                "summary": "根据 CaseID 创建部门建议",
                "parameters": [
                    {
                        "description": "提交表单",
                        "name": "case",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.Suggestion"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/vo.CommonData"
                        }
                    },
                    "401": {
                        "description": "没有认证",
                        "schema": {
                            "$ref": "#/definitions/vo.Common"
                        }
                    },
                    "422": {
                        "description": "绑定失败",
                        "schema": {
                            "$ref": "#/definitions/vo.Common"
                        }
                    }
                }
            }
        },
        "/api/v1/suggestion/delete/:id": {
            "get": {
                "description": "根据部门建议 的id号码删除，只有管理员、部门人员可以操作",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "labor",
                    "suggestion"
                ],
                "summary": "根据部门建议 的id号码删除",
                "parameters": [
                    {
                        "type": "number",
                        "description": "表单id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "正常业务处理",
                        "schema": {
                            "$ref": "#/definitions/vo.CommonData"
                        }
                    },
                    "401": {
                        "description": "未验证",
                        "schema": {
                            "$ref": "#/definitions/vo.Common"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "vo.Common": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "unauthorized"
                },
                "status": {
                    "type": "integer",
                    "example": 401
                },
                "timestamp": {
                    "type": "string",
                    "example": "2048-05-06 12:34:56"
                }
            }
        },
        "vo.CommonData": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string",
                    "example": "unauthorized"
                },
                "status": {
                    "type": "integer",
                    "example": 401
                },
                "timestamp": {
                    "type": "string",
                    "example": "2048-05-06 12:34:56"
                }
            }
        },
        "vo.DepartmentVO": {
            "type": "object",
            "properties": {
                "contact": {
                    "description": "机构联系方式",
                    "type": "string"
                },
                "description": {
                    "description": "机构介绍",
                    "type": "string"
                },
                "id": {
                    "description": "id",
                    "type": "integer"
                },
                "name": {
                    "description": "机构、单位名称",
                    "type": "string"
                },
                "service": {
                    "description": "机构提供的服务",
                    "type": "string"
                }
            }
        },
        "vo.Employer": {
            "type": "object",
            "required": [
                "employer_address",
                "employer_contact",
                "employer_faren",
                "employer_name",
                "employer_uscc"
            ],
            "properties": {
                "employer_address": {
                    "description": "地址",
                    "type": "string"
                },
                "employer_contact": {
                    "description": "联系方式",
                    "type": "string"
                },
                "employer_faren": {
                    "description": "法人",
                    "type": "string"
                },
                "employer_name": {
                    "description": "公司名",
                    "type": "string"
                },
                "employer_uscc": {
                    "description": "识别号",
                    "type": "string"
                }
            }
        },
...
        "vo.LoginResult": {
            "type": "object",
            "properties": {
                "expire": {
                    "type": "string"
                },
                "message": {
                    "type": "string",
                    "example": "unauthorized"
                },
                "refresh_url": {
                    "type": "string",
                    "example": "/api/auth/refreshToken"
                },
                "status": {
                    "type": "integer",
                    "example": 401
                },
                "timestamp": {
                    "type": "string",
                    "example": "2048-05-06 12:34:56"
                },
                "token": {
                    "type": "string",
                    "example": "asdsadasdasdasd"
                }
            }
        },
...
        "vo.Record": {
            "type": "object",
            "required": [
                "case_id",
                "name",
                "path"
            ],
            "properties": {
                "case_id": {
                    "description": "案件号，37xxx开头",
                    "type": "string"
                },
                "department_info": {
                    "description": "部门信息，新建不需要填写",
                    "type": "object",
                    "$ref": "#/definitions/vo.DepartmentVO"
                },
                "id": {
                    "description": "ID 新建不需要填写",
                    "type": "integer"
                },
                "name": {
                    "description": "记录介绍",
                    "type": "string"
                },
                "path": {
                    "description": "截图/材料等地址",
                    "type": "string"
                },
                "submitter": {
                    "description": "录入人, 新建不需要填写",
                    "type": "string"
                },
                "submitter_phone": {
                    "description": "录入人电话，新建不需要填写",
                    "type": "string"
                }
            }
        },
...
        "vo.Suggestion": {
            "type": "object",
            "required": [
                "case_id",
                "content"
            ],
            "properties": {
                "case_id": {
                    "description": "案件号，37xxx开头",
                    "type": "string"
                },
                "content": {
                    "description": "意见",
                    "type": "string"
                },
                "department": {
                    "description": "部门名称，新建不需要填写",
                    "type": "string"
                },
                "department_info": {
                    "description": "部门信息，新建不需要填写",
                    "type": "object",
                    "$ref": "#/definitions/vo.DepartmentVO"
                },
                "id": {
                    "description": "ID，新建不需要填写",
                    "type": "integer"
                },
                "submitter": {
                    "description": "录入人，新建不需要填写",
                    "type": "string"
                },
                "submitter_phone": {
                    "description": "录入人电话，新建不需要填写",
                    "type": "string"
                }
            }
        },
...
        "vo.UserRegisterEmployerForm": {
            "type": "object",
            "required": [
                "confirm_password",
                "email",
                "employer",
                "password",
                "phone"
            ],
            "properties": {
                "confirm_password": {
                    "type": "string"
                },
                "email": {
                    "type": "string",
                    "example": "test@1234.com"
                },
                "employer": {
                    "type": "object",
                    "$ref": "#/definitions/vo.Employer"
                },
                "password": {
                    "description": "Username        string   `json:\"username\" form:\"username\" binding:\"required,min=3,max=20\"`",
                    "type": "string"
                },
                "phone": {
                    "type": "string",
                    "example": "13800138000"
                }
            }
        },
...
    }
}